pyproject.toml
[project]
name = "geometric-heterogeneous-cryoem-reconstruction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lap>=0.5.12",
    "mdtraj>=1.10.2",
    "numpy>=2.2.2",
    "pillow>=11.1.0",
    "plotly>=5.24.1",
    "pytorch-lightning>=2.5.0.post0",
    "scipy>=1.15.1",
    "torchtyping>=0.1.5",
    "tqdm>=4.67.1",
    "wandb>=0.19.4",
]
